refactoring: style files cleaned
