+++ src/markup/Header/Header.tsx
@@ -1,10 +1,9 @@
 import './Header.css';
-import './meadia.css'
 import HeaderUI from "./HeaderUI/HeaderUI";
 import HeaderContent from "./HeaderContent/HeaderContent";
 
 const Header = () => {
-	return <header className="header-wrapper">
+	return <header className="header-wrapper transition-1000ms">
 			<HeaderUI />
 			{/*<HeaderContent />*/}
 	</header>
