+++ src/markup/Header/HeaderUI/HeaderUI.css
@@ -1,15 +1,15 @@
 .header-ui {
   display: grid;
   grid-template-columns: 1fr 2fr 1fr;
-  height: 2em;
   justify-content: space-between;
   align-items: center;
+  font-size: 2em;
 }
 
 .header-logo {
   display: flex;
-  width: 2em;
-  height: 2em;
+  width: 1em;
+  height: 1em;
 }
 
 .navlink {
