+++ src/markup/Header/HeaderUI/HeaderUIMedia.css
@@ -1,19 +1,10 @@
 @media screen and (max-width: 767px) {
   .header-ui {
-    font-size: 2em;
     position: fixed;
-    display: grid;
-    grid-template-columns: 1fr 2fr 1fr;
-    justify-content: space-between;
-    align-items: center;
     width: 100vw;
     top: 0;
     left: 0;
     right: 0;
-  }
-  .header-logo {
-    width: 1em;
-    height: 1em;
   }
   .header-inline-menu {
     display: none;
